feat: add docid to _id map into memory (#690)
* support index force merge with specific partition_id

* add param "enable_id_cache" to set if enable add docid to _id value map

* add docid to _id map into memory

+++ internal/engine/storage/storage_manager.h
@@ -21,7 +21,7 @@ class StorageManager {
  public:
   StorageManager(const std::string &root_path);
   ~StorageManager();
-  Status Init(int cache_size);
+  Status Init(size_t cache_size);
 
   Status Add(int cf_id, int64_t id, const uint8_t *value, int len);
